moving FibonacciBinder.kt, AnswerFibScreen.kt and FibScreen.kt files from the `presentation` module to the `features:fibonacci:impl` module
